requirements.txt
django-social-auth==0.7.22
httplib2==0.8
oauth2==1.5.211
python-openid==2.2.5
